Correções Admin
Corrigido rotas e componentes do front

+++ template/Spruha-React/src/components/ECommerce/ECDashboard/administrador/Desempenho/AVPR/ProgressoAVPR.js
@@ -2,31 +2,23 @@ import { Divider } from "@mui/material";
 import React, { Fragment, useContext, useEffect, useState } from "react";
 import { Breadcrumb, Button, Col, Row, Card, ProgressBar, Table } from 'react-bootstrap';
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
-import { usuarioContext } from "../../../../../..";
+import { deleteQuestionAlert, deleteSucessAlert } from "../../../Components/Alerts"
 import api from "../../../../../../api";
 
-import user1 from "../../../../../../assets/img/users/1.jpg";
 
 const ProgressoOKR = () => {
 
   const dadosrota = useLocation();
-
   const navegar = useNavigate();
-
   const [values, setValues] = useState({});
 
   useEffect(() => {
-
 
     setValues(dadosrota.state);
 
   }, [dadosrota.state]);
 
   // console.log(values)
-
-
-
-
 
   return (
     <Fragment>
@@ -59,40 +51,42 @@ const ProgressoOKR = () => {
 
           </div>
         </div>
-        <div className="d-flex">
-          <div className="justify-content-center">
-
-            <Button
-              variant="primary"
-              type="button"
-              className="my-2 me-2 btn-icon"
-              onClick={() => { 
-                // navegar(`${process.env.PUBLIC_URL}/avaliacao_por_resultado/`, { state: values })
-                api.delete(`avpr/delete?id=${values?.AVPRselect?.id}`).then(r=>{
-                  if(r.data.status){
-                    alert("apagado")
-                     navegar(`${process.env.PUBLIC_URL}/avaliacao_por_resultado/`, { state: values })
-                  }
-                })
-               }}
-            >
-               <i class="bi bi-file-earmark-excel"></i>
-            </Button>
-
-          </div>
-        </div>
 
       </div>
       {/* <!-- End Page Header --> */}
 
       <div className="card custom-card">
 
-        <div className="card-header border-bottom-0 d-flex flex-column">
+        <Card.Header className="border-bottom-0 d-flex justify-content-between">
 
-          <label className="main-content-label my-auto pt-2">Avaliado: </label>
-          <label className="main-content-label my-auto pt-2">Ciclo de avaliação: </label>
+          <div className="d-flex flex-column">
+            <label className="main-content-label my-auto pt-2">Avaliado: </label>
+            <label className="main-content-label my-auto pt-2">Ciclo de avaliação: </label>
+          </div>
 
-        </div>
+          <div>
+            <Button
+              variant="danger"
+              type="button"
+              className="me-2 btn-icon"
+              onClick={() => {
+                deleteQuestionAlert().then((result) => {
+                  if (result.isConfirmed) {
+                    api.delete(`avpr/delete?id=${values?.AVPRselect?.id}`).then(r => {
+                      if (r.data.status) {
+                        deleteSucessAlert()
+                        navegar(`${process.env.PUBLIC_URL}/avaliacao_por_resultado/`, { state: values })
+                      }
+                    })
+                  }
+                })
+              }}
+            >
+              <i class="bi bi-trash2-fill"></i>
+            </Button>
+          </div>
+
+        </Card.Header>
 
         <div className="card-body pt-2 pb-0">
           <div className="table-responsive tasks">
